--- a/src/jmh/java/org/javarosa/benchmarks/core/model/DeleteRepeatDagBenchmark.java
+++ b/src/jmh/java/org/javarosa/benchmarks/core/model/DeleteRepeatDagBenchmark.java
@@ -2,16 +2,10 @@ package org.javarosa.benchmarks.core.model;
 
 import static java.util.stream.IntStream.range;
 import static org.javarosa.benchmarks.BenchmarkUtils.dryRun;
-import static org.javarosa.core.util.BindBuilderXFormsElement.bind;
-import static org.javarosa.core.util.XFormsElement.body;
-import static org.javarosa.core.util.XFormsElement.head;
-import static org.javarosa.core.util.XFormsElement.html;
-import static org.javarosa.core.util.XFormsElement.input;
-import static org.javarosa.core.util.XFormsElement.mainInstance;
-import static org.javarosa.core.util.XFormsElement.model;
-import static org.javarosa.core.util.XFormsElement.repeat;
-import static org.javarosa.core.util.XFormsElement.t;
-import static org.javarosa.core.util.XFormsElement.title;
+import static org.javarosa.benchmarks.core.model.CreateRepeatDagBenchmark.getExpressionInsideScenario;
+import static org.javarosa.benchmarks.core.model.CreateRepeatDagBenchmark.getExpressionInsideWithPositionCallScenario;
+import static org.javarosa.benchmarks.core.model.CreateRepeatDagBenchmark.getExpressionInsideWithRefOutsideScenario;
+import static org.javarosa.benchmarks.core.model.CreateRepeatDagBenchmark.getSumExpressionOutsideScenario;
 
 import java.io.IOException;
 import org.javarosa.core.test.Scenario;
@@ -99,93 +93,5 @@ public class DeleteRepeatDagBenchmark {
     // Expect throughput to go down as repeat instance count goes up because position is recomputed for every instance.
     public void deleteRepeat_withPositionExpression(ExecutionPlan plan, Blackhole bh) {
         range(0, plan.repeatCount).forEach(n -> bh.consume(plan.expressionInsideWithPositionCallScenario.removeRepeat("/data/repeat[0]")));
-    }
-
-    private static Scenario getExpressionInsideScenario() throws IOException {
-        return Scenario.init("Repeat with expression inside", html(
-            head(
-                title("Repeat with expression inside"),
-                model(
-                    mainInstance(t("data id=\"sibling-expression\"",
-                        t("random"),
-
-                        t("repeat jr:template=\"\"",
-                            t("random"),
-                            t("random_x2")
-                        ))),
-                    bind("/data/random").type("int").calculate("random()"),
-                    bind("/data/repeat/random").type("int").calculate("random()"),
-                    bind("/data/repeat/random_x2").type("int").calculate("../random * 2"))),
-            body(
-                repeat("/data/repeat",
-                    input("/data/repeat/random")
-                )
-            )));
-    }
-
-    private static Scenario getExpressionInsideWithRefOutsideScenario() throws IOException {
-        return Scenario.init("Repeat with expression inside referencing outside", html(
-            head(
-                title("Repeat with expression inside referencing outside"),
-                model(
-                    mainInstance(t("data id=\"outside-ref\"",
-                        t("random"),
-
-                        t("repeat jr:template=\"\"",
-                            t("random"),
-                            t("random_outer")
-                        ))),
-                    bind("/data/random").type("int").calculate("random()"),
-                    bind("/data/repeat/random").type("int").calculate("random()"),
-                    bind("/data/repeat/random_outer").type("int").calculate("/data/random"))),
-            body(
-                repeat("/data/repeat",
-                    input("/data/repeat/random")
-                )
-            )));
-    }
-
-    private static Scenario getSumExpressionOutsideScenario() throws IOException {
-        return Scenario.init("Repeat with sum expression outside", html(
-            head(
-                title("Repeat with sum expression outside"),
-                model(
-                    mainInstance(t("data id=\"outside-sum\"",
-                        t("sum"),
-
-                        t("repeat jr:template=\"\"",
-                            t("random"),
-                            t("random_outer")
-                        ))),
-                    bind("/data/sum").type("int").calculate("sum(/data/repeat/random)"),
-                    bind("/data/repeat/random").type("int").calculate("random()"),
-                    bind("/data/repeat/random_x2").type("int").calculate("/data/repeat/random * 2"))),
-            body(
-                repeat("/data/repeat",
-                    input("/data/repeat/random")
-                )
-            )));
-    }
-
-    private static Scenario getExpressionInsideWithPositionCallScenario() throws IOException {
-        return Scenario.init("Repeat with expression inside referencing outside", html(
-            head(
-                title("Repeat with expression inside referencing outside"),
-                model(
-                    mainInstance(t("data id=\"outside-ref\"",
-                        t("random"),
-
-                        t("repeat jr:template=\"\"",
-                            t("random"),
-                            t("position")
-                        ))),
-                    bind("/data/random").type("int").calculate("random()"),
-                    bind("/data/repeat/random").type("int").calculate("random()"),
-                    bind("/data/repeat/position").type("int").calculate("position(..)"))),
-            body(
-                repeat("/data/repeat",
-                    input("/data/repeat/random")
-                )
-            )));
     }
 }