Closes #587

I also discovered that `contextualize` had another issue around when it applies multiplicities and fixed it in 6555986.

#### What has been done to verify that this works as intended?
Wrote and ran tests and benchmarks.

Benchmark results show, as expected, that there are two significant changes in performance:
- now `position(..)` calls get updated for every repeat instance. This slows down repeat creation. The tradeoff is that `count` and `sum` calls in the repeat are correctly evaluated and updated.
- repeat deletion is much faster across the board.

```
Master as of 6/25 on H’s computer
Benchmark                                  (repeatCount)  Mode     Score    Error  Units
createRepeat_withExpressionReferencingOuterNode      250  thrpt   47.069 ±  6.145  ops/s
createRepeat_withExpressionReferencingOuterNode      500  thrpt   20.433 ±  2.529  ops/s
createRepeat_withExpressionReferencingOuterNode     1000  thrpt    9.044 ±  2.759  ops/s

createRepeat_withExpressionReferencingSibling        250  thrpt   26.240 ± 18.184  ops/s
createRepeat_withExpressionReferencingSibling        500  thrpt   13.870 ±  6.945  ops/s
createRepeat_withExpressionReferencingSibling       1000  thrpt    5.976 ±  0.717  ops/s

createRepeat_withPositionExpression                  250  thrpt    8.922 ±  1.681  ops/s
createRepeat_withPositionExpression                  500  thrpt    1.728 ±  0.270  ops/s
createRepeat_withPositionExpression                 1000  thrpt    0.317 ±  0.016  ops/s

createRepeat_withSumExpressionOutside                250  thrpt   18.597 ±  5.898  ops/s
createRepeat_withSumExpressionOutside                500  thrpt    4.941 ±  0.413  ops/s
createRepeat_withSumExpressionOutside               1000  thrpt    0.929 ±  0.099  ops/s

deleteRepeat_withExpressionReferencingOuterNode      250  thrpt   95.249 ± 24.186  ops/s
deleteRepeat_withExpressionReferencingOuterNode      500  thrpt   51.091 ± 24.554  ops/s
deleteRepeat_withExpressionReferencingOuterNode     1000  thrpt   22.955 ±  2.865  ops/s

deleteRepeat_withExpressionReferencingSibling        250  thrpt   95.435 ± 22.241  ops/s
deleteRepeat_withExpressionReferencingSibling        500  thrpt   48.241 ± 22.993  ops/s
deleteRepeat_withExpressionReferencingSibling       1000  thrpt   21.860 ±  1.900  ops/s

deleteRepeat_withPositionExpression                  250  thrpt    8.070 ±  1.361  ops/s
deleteRepeat_withPositionExpression                  500  thrpt    1.701 ±  0.269  ops/s
deleteRepeat_withPositionExpression                 1000  thrpt    0.295 ±  0.009  ops/s

deleteRepeat_withSumExpressionOutside                250  thrpt   14.653 ±  1.506  ops/s
deleteRepeat_withSumExpressionOutside                500  thrpt    4.152 ±  0.294  ops/s
deleteRepeat_withSumExpressionOutside               1000  thrpt    0.801 ±  0.038  ops/s

v2.17.0 as of 6/25 on H’s computer
Benchmark                                (repeatCount)   Mode   Score   Error  Units
createRepeat_withExpressionReferencingOuterNode    250  thrpt  37.173 ± 5.824  ops/s
createRepeat_withExpressionReferencingOuterNode    500  thrpt  16.601 ± 3.566  ops/s
createRepeat_withExpressionReferencingOuterNode   1000  thrpt   6.969 ± 1.139  ops/s

createRepeat_withExpressionReferencingSibling      250  thrpt  28.540 ± 3.437  ops/s
createRepeat_withExpressionReferencingSibling      500  thrpt  13.540 ± 2.007  ops/s
createRepeat_withExpressionReferencingSibling     1000  thrpt   5.504 ± 0.599  ops/s

createRepeat_withPositionExpression                250  thrpt  25.239 ± 3.719  ops/s
createRepeat_withPositionExpression                500  thrpt  12.387 ± 1.312  ops/s
createRepeat_withPositionExpression               1000  thrpt   5.226 ± 1.372  ops/s

createRepeat_withSumExpressionOutside              250  thrpt  14.837 ± 0.653  ops/s
createRepeat_withSumExpressionOutside              500  thrpt   3.797 ± 0.214  ops/s
createRepeat_withSumExpressionOutside             1000  thrpt   0.716 ± 0.086  ops/s

deleteRepeat_withExpressionReferencingOuterNode    250  thrpt  11.573 ± 0.731  ops/s
deleteRepeat_withExpressionReferencingOuterNode    500  thrpt   2.281 ± 0.206  ops/s
deleteRepeat_withExpressionReferencingOuterNode   1000  thrpt   0.498 ± 0.034  ops/s

deleteRepeat_withExpressionReferencingSibling      250  thrpt  10.692 ± 1.810  ops/s
deleteRepeat_withExpressionReferencingSibling      500  thrpt   2.411 ± 0.251  ops/s
deleteRepeat_withExpressionReferencingSibling     1000  thrpt   0.526 ± 0.047  ops/s

deleteRepeat_withPositionExpression                250  thrpt   3.471 ± 0.833  ops/s
deleteRepeat_withPositionExpression                500  thrpt   0.842 ± 0.088  ops/s
deleteRepeat_withPositionExpression               1000  thrpt   0.149 ± 0.006  ops/s

deleteRepeat_withSumExpressionOutside              250  thrpt   6.481 ± 0.562  ops/s
deleteRepeat_withSumExpressionOutside              500  thrpt   1.399 ± 0.255  ops/s
deleteRepeat_withSumExpressionOutside             1000  thrpt   0.304 ± 0.055  ops/s
```

#### Why is this the best possible solution? Were any other approaches considered?
I tried to keep prose minimal and refer out to tests and benchmarks.

For the benchmarks, I focused on common cases and having a baseline. There's probably a better way to parameterize the cases across `Scenario`s but I couldn't quickly come up with one. I ran all cases with 250, 500, 1000 repeat instances so that we can get a sense of how the performance is affected as repeat instance count increases. I reduced warmup and measurement iterations because I got stable enough results that show relationships without the benchmarks taking too long. We now have an understanding of the complexity of each of these cases so the benchmarks are really just a sanity check.

#### How does this change affect users? Describe intentional changes to behavior and behavior that could have accidentally been affected by code changes. In other words, what are the regression risks?
The only regression risk are around `contextualize`. This only applies to repeats. I think our tests really reduce the risk.

#### Do we need any specific form for testing your changes? If so, please attach one.

#### Does this change require updates to documentation? If so, please file an issue [here]( https://github.com/getodk/docs/issues/new) and include the link below.
This is an update to documentation!