@ggalmazor 2020-06-26
Got it. I will have cycles for this during the weekend ;)