@lognaturel 2020-06-30
Thanks, @ggalmazor! There's no big rush so I'll leave open for another week and merge then. If nothing made you go 🤯🤯🤯 then it's probably good enough for now. I think you probably would have caught a major blunder even in a quick pass.