@lognaturel 2020-07-06
Thank you!

I changed "JavaRosa casts expression output values to types other than strings before committing them to the instance" to "JavaRosa's internal representation of expression results is typed. The conversion to string values happens on serialization to XML." That seems more precise to me. Does it miss something that you think is important or did I misunderstand?

I don't think "JavaRosa doesn't update the internal state of form controls associated with nodes when evaluating these expressions" is quite accurate. `TreeElement` has flags for `required`, `relevant` and `read-only` that do get updated. Again, let me know if I misunderstood. Prior to https://github.com/GetODK/javarosa/pull/581, the `relevant` flag wasn't used directly but there's reasoning there for why all the re-computations were redundant.