@ggalmazor 2020-07-06
Thanks For improving the first sentence. I agree with you that it is more
accurate now.

About the second, I know what you say about TreeElement, but I’m referring
to QuestionControl. That's why I think it’s relevant. While I was exploring
the code I was biased by how TreeElement works and then I discovered that
QuestionControl doesn’t do the same thing. This is something important to
know because some concerns, like choices in a select, and their state, are
shared among TreeElements and QuestionControl.

El El lun, 6 jul 2020 a las 3:46, Hélène Martin <notifications@github.com>
escribió:

> Thank you!
>
> I changed "JavaRosa casts expression output values to types other than
> strings before committing them to the instance" to "JavaRosa's internal
> representation of expression results is typed. The conversion to string
> values happens on serialization to XML." That seems more precise to me.
> Does it miss something that you think is important or did I misunderstand?
>
> I don't think "JavaRosa doesn't update the internal state of form controls
> associated with nodes when evaluating these expressions" is quite accurate.
> TreeElement has flags for required, relevant and read-only that do get
> updated. Again, let me know if I misunderstood. Prior to #581
> <https://github.com/GetODK/javarosa/pull/581>, the relevant flag wasn't
> used directly but there's reasoning there for why all the re-computations
> were redundant.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/GetODK/javarosa/pull/590#issuecomment-653974454>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AABSIWLMDWNBVXUSMGOQOMDR2EUHLANCNFSM4OIYUI6A>
> .
>
-- 
Guillermo Gutiérrez
@ggalmazor <http://twitter.com/ggalmazor>