@lognaturel 2020-07-06
> some concerns, like choices in a select, and their state, are shared among TreeElements and QuestionControl.

I think you mean `QuestionDef` and I see what you mean. This leads me to notice that there's nothing about the fact that dynamic selects exist entirely outside of the DAG. I have added some more context in 2d3843f and am interested to see what you think. It doesn't fit perfectly there but I think it's all important to capture. What I'd like to do eventually is add more elaborate Javadoc comments to the major classes like I did for `FormIndex`. Once that happens I could refer to those comments here.